+++ src/main/kotlin/com/nylas/models/CreateNylasListRequest.kt
@@ -22,6 +22,12 @@ data class CreateNylasListRequest(
   @Json(name = "description")
   val description: String? = null,
 ) {
+  init {
+    require(name.isNotBlank() && name.length <= 256) {
+      "name must be between 1 and 256 characters"
+    }
+  }
+
   /**
    * Builder for [CreateNylasListRequest].
    * @param name Name of the list.